fix: 코드 리뷰 사항 반영, 다중 트랜잭션 환경에서 DB 커넥션 최적화 작업
